Update cloud-exploit.sh
First, we sent a request to the form which was vulnerable by using a non-number POST request to save the 
response in the response variable declared. Following this, we extract the data from the error message, 
which contained the sensitive information like secret key, etc, by using the grep command. Now we need to 
configure AWS using these credentials we found, and then we list the files and directories using the ls s3 
command. This shows us that the environment variables which were dumped locally, can easily be found in 
the error message and then these can be used to list the buckets which contain other important information. 
Finally, we unset all the variables we set up, so as to not leave any trace in the system or cloud environment 
and clean our marks.