Create recon-automation.sh
The above script automates the process for the nmap and netcat commands we used previously. The target 
is pre-defined on this script to scan the metasploitable VM. It can also be altered to take user input. The port 
range has been predefined in this case too but can be changed in the script as needed. The script first runs 
an nmap scan on 10.0.2.5 as the target, and then to follow with it, it runs the netcat scan with the command 
nc -v -n -z -w1 10.0.2.5 100-1000 to specify details and port numbers to be scanned. The nmap runs the 
command, and echo prints our result according to the nmap_result function.  Similarly, the netcat program 
runs with the command specified and then prints the results obtained. This is very simple code to run 
automation for nmap and netcat and can be altered as per the commands required and more arguments 
can be defined as per the need.