Create web-exploit.py
This is a very simple script to execute the reflected XSS attack. The import requests import the library which 
is used to send the HTTP requests. The next line allows us to set the target URL we think may be vulnerable 
to an XSS (the attacking website). The payload variable defines the JavaScript which will be embedded in the 
user input. Following this, we need to craft the URL in a way that the payload will get added into the target 
websites’ URL, thus that variable has been defined.  Finally, the last part is to initiate the attack for which we 
will send a get request which triggers the XSS attack and then checks if that target was vulnerable or not to 
the attack. The if-else loops signify that with the print statements. If the attack has been executed, the user 
will see a pop-up that says, “You have been hacked!” with an OK button on their screen.