Create endpoint-security-detection.sql
First, we select the source category as the log source we want to use in this query. Now in this case, we want 
to know if any process has been created  and thus, we use the eventid = 4688 which stands for new process 
creation.  The next line filters the process creation events where the process command line has common 
exploit files names like .exe, .bat, .cmd. The last line again filters for events that contain SSH  or PowerShell 
command like activity which could show the process creation events leading to exploitation.  
So overall, if we look at this query, it looks for process creation events, and then filters them in windows 
defender logs for the file types that exploits usually use and then filters it finally on the basis of the files types 
that run exploits.