Create web-application-detection.sql
First, we select the source category as the log source we want to use in this query. Now, to get all the HTTP 
header details, we need to look at the raw log source. In the second line, a new field called raw_log is created, 
and the complete raw log message is extracted. The next line looks for the case-insensitive part in content 
length with (?i) and gathers the numeric value into a new field called header_size. Finally, as per the 
vulnerability, the attack occurs on POST requests, so we use the where clause to look for POST requests in 
the “method” field and filter on the basis of header size > 500 or status_code = 413. We want to make sure 
that we get all POST requests that may have an unexpectedly big header size or a status code of 413 to 
denote a large payload.